+++ src/functions/fetch_stores/structs.rs
