+++ src/functions/logs/structs.rs
