+++ src/routes/user_control.rs
@@ -1,8 +1,8 @@
-use crate::signing::decode_token_data;
-use crate::user_control::structs::*;
-use crate::user_control::*;
+use crate::functions::signing::decode_token_data;
+use crate::functions::user_control::structs::*;
+use crate::functions::user_control::*;
 use crate::utils::structs::APIErrors;
-use crate::request_guard::api_key::ApiKey;
+use crate::server::request_guard::api_key::ApiKey;
 
 use oracle::pool::Pool;
