+++ src/routes/version_check.rs
@@ -4,8 +4,8 @@ use rocket::serde::json::Json;
 use rocket::{post, State};
 
 
-use crate::version_check::get_latest_version;
-use crate::version_check::structs::*;
+use crate::functions::version_check::get_latest_version;
+use crate::functions::version_check::structs::*;
 
 #[post("/version", data = "<params>")]
 pub async fn route_version_check(
