+++ src/server/fairings/cors.rs
